tick: Introduce hrtimer based broadcast
On some architectures, in certain CPU deep idle states the local timers
stop. An external clock device is used to wakeup these CPUs. The kernel
support for the wakeup of these CPUs is provided by the tick broadcast
framework by using the external clock device as the wakeup source.

However not all implementations of architectures provide such an
external clock device. This patch includes support in the broadcast
framework to handle the wakeup of the CPUs in deep idle states on such
systems by queuing a hrtimer on one of the CPUs, which is meant to
handle the wakeup of CPUs in deep idle states.

This patchset introduces a pseudo clock device which can be registered
by the archs as tick_broadcast_device in the absence of a real external
clock device. Once registered, the broadcast framework will work as is
for these architectures as long as the archs take care of the
BROADCAST_ENTER notification failing for one of the CPUs. This CPU is
made the stand by CPU to handle wakeup of the CPUs in deep idle and it
*must not enter deep idle states*.

The CPU with the earliest wakeup is chosen to be this CPU. Hence this
way the stand by CPU dynamically moves around and so does the hrtimer
which is queued to trigger at the next earliest wakeup time. This is
consistent with the case where an external clock device is present. The
smp affinity of this clock device is set to the CPU with the earliest
wakeup. This patchset handles the hotplug of the stand by CPU as well by
moving the hrtimer on to the CPU handling the CPU_DEAD notification.

Originally-from: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>
Cc: deepthi@linux.vnet.ibm.com
Cc: paulmck@linux.vnet.ibm.com
Cc: fweisbec@gmail.com
Cc: paulus@samba.org
Cc: srivatsa.bhat@linux.vnet.ibm.com
Cc: svaidy@linux.vnet.ibm.com
Cc: peterz@infradead.org
Cc: benh@kernel.crashing.org
Cc: rafael.j.wysocki@intel.com
Cc: linuxppc-dev@lists.ozlabs.org
Link:http://lkml.kernel.org/r/20140207080632.17187.80532.stgit@preeti.in.ibm.com
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>