msm_thermal: queue work on system_power_efficient_wq
msm_thermal simple driver uses workqueues and it has no real dependency of scheduling these on the
cpu which scheduled them.

On a idle system, it is observed that and idle cpu wakes up many times just to
service this work. It would be better if we can schedule it on a cpu which the
scheduler believes to be the most appropriate one.

This patch replaces system_wq with system_power_efficient_wq

Signed-off-by: Joshua Primero <jprimero155@gmail.com>